requirements.txt
streamlit
openai
langchain
faiss-cpu
PyMuPDF
pandas
plotly
tqdm
python-dotenv
azure-ai-inference
azure-core

